docs: expand and clarify usage instructions in gpt-doc.el
Expanded and clarified on the instructions on how to use the gpt-doc in
gpt-doc.el. This includes detailing how to document variable,
function, and macro definitions both synchronously and by streaming
response. The explanation of how to include related definitions with
prefix arguments has also been updated. Furthermore, it has been
specified that gpt-doc-document-current-function generates documentation
for the current function in the buffer synchronously, and that the curl
request can be aborted with command gpt-doc-stream-abort.